+++ libft/ft_putnbr.c
@@ -6,13 +6,13 @@
 /*   By: ckrommen <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/27 21:37:06 by ckrommen          #+#    #+#             */
-/*   Updated: 2017/09/27 21:48:23 by ckrommen         ###   ########.fr       */
+/*   Updated: 2018/02/09 15:03:20 by lprior           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(long long int n)
 {
 	if (n == -2147483648)
 		ft_putstr("-2147483648");
